feature(5-1): грубый перенос логики в презентер
